Move token swap check to get_token_swap_instance
