+++ anker/src/state.rs
@@ -1,5 +1,4 @@
-use crate::instruction::{ChangeTokenSwapPoolAccountsInfo, SellRewardsAccountsInfo};
-use crate::logic::get_token_swap_instance;
+use crate::instruction::SellRewardsAccountsInfo;
 use crate::{
     error::AnkerError, ANKER_MINT_AUTHORITY, ANKER_RESERVE_AUTHORITY, ANKER_STSOL_RESERVE_ACCOUNT,
     ANKER_UST_RESERVE_ACCOUNT,
@@ -9,9 +8,11 @@ use lido::state::Lido;
 use lido::token::{Rational, StLamports};
 use lido::util::serialize_b58;
 use serde::Serialize;
+use solana_program::program_error::ProgramError;
 use solana_program::{
     account_info::AccountInfo, entrypoint::ProgramResult, msg, program_pack::Pack, pubkey::Pubkey,
 };
+use spl_token_swap::state::SwapV1;
 
 use crate::token::{self, BLamports};
 
@@ -272,21 +273,49 @@ impl Anker {
         Anker::check_is_spl_token_account("Solido's stSOL", &solido.st_sol_mint, token_account_info)
     }
 
+    /// Get an instance of the Token Swap V1 from the provided account info.
+    pub fn get_token_swap_instance(
+        &self,
+        token_swap_account: &AccountInfo,
+    ) -> Result<spl_token_swap::state::SwapV1, ProgramError> {
+        self.check_token_swap_pool(token_swap_account)?;
+        // We do not check the owner of the `token_swap_account`. Since we store
+        // this address in Anker's state, and we also trust the manager that changes
+        // this address, we don't verify the account's owner. This also allows us to
+        // test different token swap programs ids on different clusters.
+        // Check that version byte corresponds to V1 version byte.
+        if token_swap_account.data.borrow().len() != spl_token_swap::state::SwapVersion::LATEST_LEN
+        {
+            msg!(
+                "Length of the Token Swap is invalid, expected {}, found {}",
+                spl_token_swap::state::SwapVersion::LATEST_LEN,
+                token_swap_account.data.borrow().len()
+            );
+            return Err(AnkerError::WrongSplTokenSwapParameters.into());
+        }
+        if token_swap_account.data.borrow()[0] != 1u8 {
+            msg!(
+            "Token Swap instance version is different from what we expect, expected 1, found {}",
+            token_swap_account.data.borrow()[0]
+        );
+            return Err(AnkerError::WrongSplTokenSwapParameters.into());
+        }
+        // We should ignore the version 1st byte for the unpack.
+        spl_token_swap::state::SwapV1::unpack(&token_swap_account.data.borrow()[1..])
+    }
+
     /// Check if we can change the token swap account.
     pub fn check_change_token_swap_pool(
         &self,
         solido: &Lido,
-        accounts: &ChangeTokenSwapPoolAccountsInfo,
+        current_token_swap: SwapV1,
+        new_token_swap: SwapV1,
     ) -> ProgramResult {
         // We don't check that the old pool's owner is the same as the new
         // pool's owner. It's the manager's responsibility to replace the token
         // pool swap with a valid one. This also allows us to change the pool
         // program, if necessary.
         // Check if the token swap account is the same one as the stored in the instance.
-        self.check_token_swap_pool(accounts.current_token_swap_pool)?;
-        let current_token_swap = get_token_swap_instance(accounts.current_token_swap_pool)?;
-        // Checks if the provided account is a valid Token Swap instance.
-        let new_token_swap = get_token_swap_instance(accounts.new_token_swap_pool)?;
 
         // Get stSOL and UST mint. We trust that the UST mint stored in the current instance is right.
         let (st_sol_mint, ust_mint) = if current_token_swap.token_a_mint == solido.st_sol_mint {
@@ -338,7 +367,7 @@ impl Anker {
         Ok(())
     }
 
-    pub fn check_token_swap_pool(&self, token_swap_account: &AccountInfo) -> ProgramResult {
+    fn check_token_swap_pool(&self, token_swap_account: &AccountInfo) -> ProgramResult {
         if &self.token_swap_pool != token_swap_account.key {
             msg!(
                 "Invalid Token Swap instance, expected {}, found {}",
@@ -361,10 +390,9 @@ impl Anker {
         accounts: &SellRewardsAccountsInfo,
     ) -> ProgramResult {
         // Check if the token swap account is the same one as the stored in the instance.
-        self.check_token_swap_pool(accounts.token_swap_pool)?;
-        let token_swap = get_token_swap_instance(accounts.token_swap_pool)?;
-        // Check token swap instance parameters.
+        let token_swap = self.get_token_swap_instance(accounts.token_swap_pool)?;
 
+        // Check token swap instance parameters.
         // Check UST token accounts.
         self.check_ust_reserve_address(
             anker_program_id,
